[Prototype]
This is an educational android game to teach the students the basic of C++.

Player Movement, Camera following the Player (Cinemachine plugin), Dialogue System for NPCs, Level-based System, Storing of High Scores in each level using PlayerPrefs, health stats of the player, Quiz System, Pause, randomizing the choices have been implemented in the game.
PlayerPrefs function is used to store integer, float or string in a separate file.

The User Interface of the current game used some of the Unity free assets, Sunny Land resources and Standard Assets, and self-made 2d and 3d Assets(2D sprites, 2D UI buttons and etc, 3D terrain, 3D trees). The current logo is the NPC (Consola is her name) of the game.

The music resources used in this game are free from FreeSound.org and Youtube Audio Library. Credits can be found in the main scene. A pool water texture was added in a material of a water plane.

The current game has a total of 4 scenes: Main Menu, levels selection, prototype level, and player game over scene. Player game over scene is separated from the prototype scene, because the scene doesn't need to contain the current data of the level.

2D character sprites can cast shadows on 3D objects, but it cannot receive any shadows from 2D and 3D objects. The 2D character sprites includes the main player(boy in a school uniform), Consola(NPC), and bugs(enemies).

updated the script of the NPC

Imported music resources and images for the comic. Converted all imported images to Sprite 2D and UI. Updated the UI of the Game Over Screen.

Git commit that has a unity .gitignore using the GitHub Client for Desktop